fix: basic/full wallet initialization
