+++ src/redux/sagas/setupUserContext.ts
@@ -1,30 +1,43 @@
 import { all, call, fork, put } from 'redux-saga/effects';
 
-import motionSagas from './motions';
-import { ColonyWallet, isFullWallet } from '~types';
-
-import actionsSagas from './actions';
-import colonySagas, { colonyCreateSaga } from './colony';
-import extensionSagas from './extensions';
-
-import decisionsSagas from './decisions';
-
-// import vestingSagas from './vesting';
-import { setupUsersSagas } from './users';
+import { getContext, setContext, ContextModule } from '~context';
+// import AppLoadingState from '~context/appLoadingState';
+import { ColonyWallet } from '~types';
+import { getLastWallet, LastWallet, setLastWallet } from '~utils/autoLogin';
+import { createAddress } from '~utils/web3';
 
 import { ActionTypes } from '../actionTypes';
 import { AllActions } from '../types/actions';
-
-import { setContext, ContextModule } from '~context';
-
-// import AppLoadingState from '~context/appLoadingState';
-
-import { getGasPrices, putError } from './utils';
-import setupWalletContext from './setupWalletContext';
-import getOnboard from './wallet/onboard';
-import { setupTransactionSagas } from './transactions';
+import actionsSagas from './actions';
+import colonySagas, { colonyCreateSaga } from './colony';
+import decisionsSagas from './decisions';
 import expendituresSagas from './expenditures';
+import extensionSagas from './extensions';
+import motionSagas from './motions';
 // import { setupUserBalanceListener } from './setupUserBalanceListener';
+import { setupTransactionSagas } from './transactions';
+import { disconnectWallet, setupUsersSagas } from './users';
+import { getGasPrices, putError } from './utils';
+import { getBasicWallet, getWallet } from './wallet';
+// import vestingSagas from './vesting';
+import getOnboard from './wallet/onboard';
+
+const ONBOARD_METAMASK_WALLET_LABEL = 'MetaMask';
+
+const getMetamaskAddress = () => {
+  // try/catch just in case createAddress errors
+  try {
+    if (window.ethereum) {
+      return createAddress(
+        // @ts-ignore
+        window.ethereum.selectedAddress,
+      );
+    }
+  } catch {
+    // silent
+  }
+  return undefined;
+};
 
 function* setupContextDependentSagas() {
   // const appLoadingState: typeof AppLoadingState = AppLoadingState;
@@ -48,6 +61,17 @@ function* setupContextDependentSagas() {
   ]);
 }
 
+function* initializeBasicWallet(lastWallet: LastWallet) {
+  const wallet = yield call(getBasicWallet, lastWallet);
+  setContext(ContextModule.Wallet, wallet);
+}
+
+function* initializeFullWallet(lastWallet: LastWallet | null) {
+  const wallet = yield call(getWallet, lastWallet);
+  setContext(ContextModule.Wallet, wallet);
+  yield call(getGasPrices);
+}
+
 /*
  * Given an action to get the user’s wallet, use this wallet to initialise the initial
  * context that depends on it (the wallet itself, the DDB, the ColonyManager),
@@ -61,14 +85,50 @@ export default function* setupUserContext() {
     /*
      * Get the new wallet and set it in context.
      */
-    const wallet: ColonyWallet | undefined = yield call(setupWalletContext);
+
+    let wallet: ColonyWallet | undefined;
+
+    try {
+      wallet = getContext(ContextModule.Wallet);
+
+      const selectedMetamaskAddress = getMetamaskAddress();
+      /*
+       * If the wallet we've pulled from context does not have the same address as the selected account
+       * in Metamask, it's because the user just switched their account in metamask.
+       */
+      if (
+        selectedMetamaskAddress &&
+        wallet.address !== selectedMetamaskAddress &&
+        wallet.label === ONBOARD_METAMASK_WALLET_LABEL
+      ) {
+        disconnectWallet(wallet.label); // disconnect previous wallet
+
+        // replace it in local storage with the wallet the user switched to
+        setLastWallet({
+          type: ONBOARD_METAMASK_WALLET_LABEL,
+          address: selectedMetamaskAddress,
+        });
+
+        wallet = undefined;
+      }
+    } catch {
+      // wallet not seen in context yet
+    }
+
+    const lastWallet = getLastWallet();
+
+    if (!wallet && lastWallet) {
+      // Perform quick login, then run background task to fully login
+      yield call(initializeBasicWallet, lastWallet);
+      yield fork(initializeFullWallet, lastWallet);
+    } else {
+      // Perfom full wallet login
+      yield call(initializeFullWallet, lastWallet);
+    }
+
     yield put<AllActions>({
       type: ActionTypes.WALLET_OPEN_SUCCESS,
     });
-
-    if (isFullWallet(wallet)) {
-      yield call(getGasPrices);
-    }
 
     /*
      * This needs to happen first because USER_CONTEXT_SETUP_SUCCESS causes a redirect
